@@ -6,7 +6,7 @@ Dummy::Application.configure do
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs. Don't rely on the data there!
   config.secret_key_base = ('x'*30) rescue nil
-  config.secret_token = ('x'*30) rescue nil
+  # config.secret_token = ('x'*30) rescue nil
 
   config.cache_classes = true
 
